rename controller folder to match other folders
